add loggs

+++ Seminar_11/HW_05/My001Calc/Models/DivModel.java
@@ -1,19 +1,21 @@
 package Seminar_11.HW_05.My001Calc.Models;
 
+import Seminar_11.HW_05.My001Calc.Base.CalcModel;
+
 public class DivModel extends CalcModel {
 
     @Override
-    public int result() {
+    public double result() {
         return x / y;
     }
 
     @Override
-    public void setX (int value) {
+    public void setX (double value) {
         super.x = value;    
     }
 
     @Override
-    public void setY(int value) {
+    public void setY(double value) {
         super.y = value;   
     }
